backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
pdfplumber==0.11.4
docx2txt==0.8
langchain==0.3.0
langchain-google-genai==2.0.0
python-dotenv==1.0.1
